Split triage specs into separate files
